test: place broker guard in architecture suite
